LZMADecompressor, LZMACompressor, and LZMAError live in _lzma
