+++ stdlib/VERSIONS
@@ -45,6 +45,7 @@ _io: 3.0-
 _json: 3.0-
 _locale: 3.0-
 _lsprof: 3.0-
+_lzma: 3.3-
 _markupbase: 3.0-
 _msi: 3.0-3.12
 _operator: 3.4-
